Refactor: move function to other sub-module
