+++ src/conduit/articles.rs
@@ -61,16 +61,3 @@ pub fn find_one(repo: &Repo, slug_value: &str) -> Result<(Article, User), Error>
             .first(&conn)
     })
 }
-
-/// Return the number of users who have marked a specific article as favorited.
-pub fn n_favorites(repo: &Repo, article_id_value: i32) -> Result<i64, Error> {
-    use crate::db::schema::favorites::dsl::{article_id, favorites, user_id};
-    use diesel::dsl::count;
-
-    repo.run(move |conn| {
-        favorites
-            .filter(article_id.eq(article_id_value))
-            .select(count(user_id))
-            .get_result(&conn)
-    })
-}
