+++ tests/conduit_favorites.rs
@@ -3,7 +3,6 @@ mod helpers;
 use helpers::test_db::get_test_repo;
 use helpers::{create_article, create_user};
 
-use realworld_tide::conduit::articles::n_favorites;
 use realworld_tide::conduit::favorites;
 
 #[test]
@@ -30,5 +29,5 @@ fn you_can_favorite_an_article_twice_but_it_only_counts_for_one() {
     let result = favorites::favorite(&repo, user.id, article.id);
     assert!(result.is_ok());
 
-    assert_eq!(1, n_favorites(&repo, article.id).unwrap());
+    assert_eq!(1, favorites::n_favorites(&repo, article.id).unwrap());
 }
